create a schema for selecting primary activities, refactor
